+++ src/client/v2/algod/dryrun.ts
@@ -9,7 +9,7 @@ export default class Dryrun extends JSONRequest {
 
   constructor(c: HTTPClient, dr: modelsv2.DryrunRequest) {
     super(c);
-    this.blob = encoding.encode(dr.get_obj_for_encoding());
+    this.blob = encoding.encode(dr.get_obj_for_encoding(true));
   }
 
   // eslint-disable-next-line class-methods-use-this