+++ src/encoding/uint64.ts
@@ -8,7 +8,7 @@
 export function encodeUint64(num: number | bigint) {
   const isInteger = typeof num === 'bigint' || Number.isInteger(num);
 
-  if (!isInteger || num < 0 || num > 0xffffffffffffffffn) {
+  if (!isInteger || num < 0 || num > BigInt('0xffffffffffffffff')) {
     throw new Error('Input is not a 64-bit unsigned integer');
   }
 
@@ -37,7 +37,6 @@ export function encodeUint64(num: number | bigint) {
  * @returns The integer that was encoded in the input data. The return type will
  *   be determined by the parameter decodingMode.
  */
-/* eslint-disable  no-unused-vars,no-redeclare */
 export function decodeUint64(data: Uint8Array, decodingMode: 'safe'): number;
 export function decodeUint64(
   data: Uint8Array,
@@ -45,7 +44,6 @@ export function decodeUint64(
 ): number | bigint;
 export function decodeUint64(data: Uint8Array, decodingMode: 'bigint'): bigint;
 export function decodeUint64(data: any, decodingMode: any = 'safe') {
-  /* eslint-enable  no-unused-vars,no-redeclare */
   if (
     decodingMode !== 'safe' &&
     decodingMode !== 'mixed' &&